IBKRAV-ibkrav-qa/requirements.txt
ib_insync
pandas
python-dotenv
requests
pyinstaller
matplotlib
mplfinance
ttkbootstrap
python3-pytest